1st pass rewrite of UI for manage users, admin activation now works, a=chris

+++ models/user_model.php
@@ -50,7 +50,6 @@ require_once BASE_DIR."/lib/utility.php";
 class UserModel extends Model
 {
 
-
     /**
      * Just calls the parent class constructor
      */
@@ -73,6 +72,11 @@ class UserModel extends Model
         $user_id = $this->db->escapeString($user_id);
 
         $activities = array();
+        $status = $this->getUserStatus($user_id);
+        if(!$status || in_array($status, array(BANNED_STATUS,
+            INACTIVE_STATUS))) {
+            return array();
+        }
         $locale_tag = getLocaleTag();
 
         $sql = "SELECT LOCALE_ID FROM LOCALE ".
@@ -219,11 +223,54 @@ class UserModel extends Model
      *
      *  @return array a list of user
      */
-    function getUserListGroups()
+    function getUserIdUsernameList()
     {
         $this->db->selectDB(DB_NAME);
          $sql = "SELECT U.USER_ID AS USER_ID, U.USER_NAME AS USER_NAME ".
             " FROM USER U";
+        $result = $this->db->execute($sql);
+        $i = 0;
+        while($users[$i] = $this->db->fetchArray($result)) {
+            $i++;
+        }
+        unset($users[$i]); //last one will be null
+        return $users;
+    }
+
+    /**
+     *
+     */
+    function getUsers($restrict_to_group = "", $order_by="" , $dir="",
+        $limit=0, $num=100)
+    {
+        $restrict_to_group =$this->db->escapeString($restrict_to_group);
+        $user_columns = array("USER_ID", "FIRST_NAME", "LAST_NAME",
+            "USER_NAME", "EMAIL");
+        if(!in_array($order_by, $user_columns)) {
+            $order_by = "USER_NAME";
+        }
+        if(!in_array($dir, array("ASC, DESC"))) {
+            $dir = "ASC";
+        }
+        if(!is_int($limit)) {
+            $limit = 0;
+        }
+        if(!is_int($num)) {
+            $num = 100;
+        }
+        if($restrict_to_group == "") {
+            $sql = "SELECT USER_NAME, FIRST_NAME, LAST_NAME,
+                EMAIL, STATUS FROM USER ORDER BY $order_by $dir 
+                LIMIT $limit, $num";
+        } else {
+            $sql = "SELECT  U.USER_NAME AS USER_NAME,
+                U.FIRST_NAME AS FIRST_NAME, U.LAST_NAME AS LAST_NAME,
+                U.EMAIL AS EMAIL, U.STATUS AS STATUS FROM USER U, USER_GROUP UG,
+                GROUPS G WHERE G.GROUP_NAME = '$restrict_to_group' AND
+                U.USER_ID=UG.USER_ID AND UG.GROUP_ID=G.GROUP_ID
+                ORDER BY $order_by $dir LIMIT $limit, $num";
+        }
+
         $result = $this->db->execute($sql);
         $i = 0;
         while($users[$i] = $this->db->fetchArray($result)) {
@@ -249,39 +296,18 @@ class UserModel extends Model
     }
 
     /**
-     * Add a user with a given username and password to the list of users
-     * that can login to the admin panel
+     * get a status of user by user_id
      *
-     * @param string $username  the username of the user to be added
-     * @param string $password  the password of the user to be added
-     * @param string $firstname the firstname of the user to be added
-     * @param string $lastname the lastname of the user to be added
-     * @param string $email the email of the user to be added
-     * @return bool whether the operation was successful
+     * @param string $user_id id of the user
      */
-    function addUser($username, $password, $firstname='', $lastname='',
-        $email='')
+    function getUserStatus($user_id)
     {
+        $user_id = $this->db->escapeString($user_id);
         $this->db->selectDB(DB_NAME);
-        $sql = "INSERT INTO USER(FIRST_NAME, LAST_NAME, 
-                USER_NAME, EMAIL, ACTIVE, PASSWORD, HASH) VALUES ('".
-        $this->db->escapeString($firstname)."', '".
-        $this->db->escapeString($lastname)."', '".
-        $this->db->escapeString($username)."', '".
-        $this->db->escapeString($email)."', 1,'".
-            crawlCrypt($this->db->escapeString($password))."','".
-            crawlCrypt($this->db->escapeString($username))."') ";
+        $sql = "SELECT STATUS FROM USER WHERE USER_ID=$user_id;";
         $result = $this->db->execute($sql);
-        if(!$user_id = $this->getUserId($username)) {
-            return false;
-        }
-        $sql = "INSERT INTO USER_GROUP (USER_ID, GROUP_ID) VALUES('".
-            $this->db->escapeString($user_id)."', '1')";
-        $result = $this->db->execute($sql);
-        $sql = "INSERT INTO USER_ROLE  VALUES ('".
-            $this->db->escapeString($user_id)."', 2) ";
-        $result_id = $this->db->execute($sql);
-        return true;
+        $row = $this->db->fetchArray($result);
+        return $row['STATUS'];
     }
 
     /**
@@ -303,6 +329,79 @@ class UserModel extends Model
     }
 
     /**
+     *
+     * @param string $username
+     */
+    function getUser($username)
+    {
+        $this->db->selectDB(DB_NAME);
+        $sql = "SELECT * FROM USER WHERE 
+         USER_NAME = '$username' LIMIT 1";
+        $result = $this->db->execute($sql);
+        if(!$result) {
+            return false;
+        }
+        $row = $this->db->fetchArray($result);
+        return $row;
+    }
+
+    /**
+     * get a status of user by user_id
+     *
+     * @param string $user_id id of the user
+     */
+    function updateUserStatus($user_id, $status)
+    {
+        $user_id = $this->db->escapeString($user_id);
+        if(!in_array($status, array(ACTIVE_STATUS, INACTIVE_STATUS,
+            BANNED_STATUS))) {
+            return;
+        }
+        $this->db->selectDB(DB_NAME);
+        $sql = "UPDATE USER SET STATUS=$status WHERE USER_ID=$user_id";
+        $this->db->execute($sql);
+    }
+
+    /**
+     * Add a user with a given username and password to the list of users
+     * that can login to the admin panel
+     *
+     * @param string $username  the username of the user to be added
+     * @param string $password  the password of the user to be added
+     * @param string $firstname the firstname of the user to be added
+     * @param string $lastname the lastname of the user to be added
+     * @param string $email the email of the user to be added
+     * @return bool whether the operation was successful
+     */
+    function addUser($username, $password, $firstname='', $lastname='',
+        $email='', $status = ACTIVE_STATUS)
+    {
+        $this->db->selectDB(DB_NAME);
+        $sql = "INSERT INTO USER(FIRST_NAME, LAST_NAME, 
+            USER_NAME, EMAIL, PASSWORD, STATUS, HASH) VALUES ('".
+            $this->db->escapeString($firstname)."', '".
+            $this->db->escapeString($lastname)."', '".
+            $this->db->escapeString($username)."', '".
+            $this->db->escapeString($email)."', '".
+            crawlCrypt($this->db->escapeString($password))."','".
+            $this->db->escapeString($status)."', '".
+            time().'$'.md5($row['USER_NAME'].AUTH_KEY.time())."') ";
+        $result = $this->db->execute($sql);
+        if(!$user_id = $this->getUserId($username)) {
+            return false;
+        }
+        $user_id = $this->db->escapeString($user_id);
+        $sql = "INSERT INTO USER_GROUP (USER_ID, GROUP_ID) VALUES('".
+            $user_id."', '" . PUBLIC_GROUP_ID . "')";
+        $result = $this->db->execute($sql);
+        $sql = "INSERT INTO USER_ROLE  VALUES ('". $user_id."', 2) ";
+        $result_id = $this->db->execute($sql);
+        return true;
+    }
+
+
+
+    /**
      * Deletes a user by username from the list of users that can login to
      * the admin panel
      *
@@ -311,21 +410,35 @@ class UserModel extends Model
     function deleteUser($username)
     {
         $this->db->selectDB(DB_NAME);
-        $userid = $this->getUserId($username);
+        $userid = $this->db->escapeString($this->getUserId($username));
         if($userid) {
-            $sql = "DELETE FROM USER_ROLE WHERE USER_ID='".
-                $this->db->escapeString($userid)."'";
+            $sql = "DELETE FROM USER_ROLE WHERE USER_ID='".$userid."'";
             $result = $this->db->execute($sql);
-            $sql = "DELETE FROM USER_GROUP WHERE USER_ID='".
-                $this->db->escapeString($userid)."'";
+            $sql = "DELETE FROM USER_GROUP WHERE USER_ID='".$userid."'";
             $result = $this->db->execute($sql);
         }
-        $sql = "DELETE FROM USER WHERE USER_NAME='".
-            $this->db->escapeString($username)."'";
+        $sql = "DELETE FROM USER WHERE USER_ID='".$userid."'";;
         $result = $this->db->execute($sql);
     }
 
-
+    /**
+     *
+     */
+    function updateUser($user)
+    {
+        $user_id = $user['USER_ID'];
+        unset($user['USER_ID']);
+        unset($user['USER_NAME']);
+        $this->db->selectDB(DB_NAME);
+        $sql = "UPDATE USER SET ";
+        $comma ="";
+        foreach($user as $field=>$value) {
+            $sql .= "$comma $field='$value' ";
+            $comma = ",";
+        }
+        $sql .= " WHERE USER_ID=$user_id";
+        $this->db->execute($sql);
+    }
     /**
      * Adds a role to a given user
      *
